Add orchestrator - Move majority of "system 2" processing to orchestrator - Remove need to run secondary app server - typing cleanup
