+++ system_two_model.py
@@ -3,7 +3,7 @@ from enum import StrEnum, auto
 import ollama
 from pydantic import BaseModel
 
-from metacognitive import MetacognitiveVector, MetacognitiveVectorComputation
+from metacognitive import MetacognitiveVector
 from prompts import PromptNames, Prompts
 from system_communication_objects import SystemTwoRequest
 
@@ -14,6 +14,7 @@ class NodeRole(StrEnum):
     Evaluator = auto()
     Generalist = auto()
     Synthesizer = auto()
+    System_One = auto()
 
 
 class NodeResponse(BaseModel):
@@ -32,6 +33,9 @@ class SystemTwoResponse(BaseModel):
 
 
 class Node:
+    def __init__(self, role: NodeRole = NodeRole.Generalist) -> None:
+        self.role = role
+
     role: NodeRole = NodeRole.Generalist
     # TODO adjust weights!
     role_weights: dict[NodeRole, dict[str, float]] = {
@@ -120,136 +124,3 @@ class Node:
 
         response = ollama.chat(model="llama3.2", messages=messages)
         return response.message.content
-
-
-class SystemTwo:
-    def __init__(self):
-        self.nodes = [Node(), Node()]
-        self.taken_roles: dict[NodeRole, Node | None] = {
-            NodeRole.Domain_Expert: None,
-            NodeRole.Critic: None,
-            NodeRole.Evaluator: None,
-            NodeRole.Generalist: None,
-            NodeRole.Synthesizer: None,
-        }
-
-    def _reset_taken_nodes(self) -> None:
-        self.taken_roles: dict[NodeRole, Node | None] = {
-            NodeRole.Domain_Expert: None,
-            NodeRole.Critic: None,
-            NodeRole.Evaluator: None,
-            NodeRole.Generalist: None,
-            NodeRole.Synthesizer: None,
-        }
-
-    def _transition_nodes(self, system_one_vector: MetacognitiveVector):
-        self._reset_taken_nodes()
-
-        for node in self.nodes:
-            # TODO: manage role balance w/ Hungarian algo, right now use first available
-            role_preferences = node.get_role_preferences(system_one_vector)
-            sorted_role_preferences = [
-                role
-                for role, _ in sorted(
-                    role_preferences.items(), key=lambda r: r[1], reverse=True
-                )
-            ]
-            for role in sorted_role_preferences:
-                if self.taken_roles[role] == None:
-                    self.taken_roles[role] = node
-                    node.assign_role(role)
-                    break
-
-    async def get_response(
-        self,
-        user_prompt: str,
-        system_one_response: str,
-        system_one_vector: MetacognitiveVector,
-        prompts: Prompts,
-        weights,
-    ) -> SystemTwoResponse:
-
-        messages = [
-            {
-                "role": "system",
-                "content": prompts.get_prompt(
-                    PromptNames.System_Two_System, context={}
-                ),
-                "thinking": "true",
-            },
-            {"role": "assistant", "content": system_one_response},
-        ]
-
-        self._transition_nodes(system_one_vector)
-
-        role_responses: list[NodeResponse] = []
-        previous_response = system_one_response
-        previous_role = "system one"
-        synthesizer_response: str | None = None
-        synthesizer_msv: MetacognitiveVector | None = None
-        for role, node in self.taken_roles.items():
-            if node:
-                node_response = node.get_response(
-                    user_prompt, previous_response, previous_role, prompts
-                )
-
-                state = await MetacognitiveVectorComputation.compute_metacognitive_state_vector(
-                    "baseline", prompts, weights, node_response, previous_response
-                )
-                role_responses.append(
-                    NodeResponse(
-                        node_role=role, node_response=node_response, node_msv=state
-                    )
-                )
-                if role == NodeRole.Synthesizer:
-                    synthesizer_response = node_response
-                    synthesizer_msv = state
-
-                previous_response = node_response
-                previous_role = role
-                messages.append({"role": "assistant", "content": node_response})
-        if not synthesizer_response:
-            messages.append(
-                {
-                    "role": "user",
-                    "content": prompts.get_prompt(
-                        PromptNames.System_Two_User,
-                        context={"user_prompt": user_prompt},
-                    ),
-                }
-            )
-
-            overall_system_two_response = ollama.chat(
-                model="llama3.2", messages=messages
-            ).message.content
-            state = (
-                await MetacognitiveVectorComputation.compute_metacognitive_state_vector(
-                    "baseline",
-                    prompts,
-                    weights,
-                    overall_system_two_response,
-                    system_one_response,
-                )
-            )
-        else:
-            overall_system_two_response = synthesizer_response
-            state = synthesizer_msv
-
-        return SystemTwoResponse(
-            node_responses=role_responses,
-            system_two_response=overall_system_two_response,
-            metacognitive_vector=state,
-        )
-
-
-system_two = SystemTwo()
-
-
-async def get_response(system_two_request: SystemTwoRequest) -> SystemTwoResponse:
-    return await system_two.get_response(
-        system_two_request.user_prompt,
-        system_two_request.system_one_response,
-        system_two_request.metacognitive_vector,
-        system_two_request.prompts,
-        system_two_request.weights,
-    )
